requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
dateparser==1.2.1
dnspython==2.7.0
dotenv==0.9.9
fastjsonschema==2.21.1
fonttools==4.57.0
frozenlist==1.6.0
idna==3.10
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.1
multidict==6.4.3
narwhals==1.38.0
nbformat==5.10.4
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
propcache==0.3.1
pycryptodome==3.22.0
pymongo==4.12.1
pyparsing==3.2.3
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-louvain==0.16
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
ta==0.11.0
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
